Fixed docker container to use pipenv; currently does not exit with ^C
